refactor(runtime): W5 — flatten session store into RuntimeSessionRepository (Doc 2 §6)
Remove session/store/ subpackage: SessionStore->RuntimeSessionRepository, InMemorySessionStore->InMemoryRuntimeSessionRepository, SessionStoreFactory->RuntimeSessionRepositoryFactory, DefaultSessionStoreFactory->DefaultRuntimeSessionRepositoryFactory, SessionNotFoundException flattened to session/. Persistence is now expressed as the RuntimeSessionRepository interface + InMemory first impl, no store/ subpackage. (Session->RuntimeSession model rename deferred.)

Verified: clean test-compile + 68 non-Mockito tests green; example e2e green.
