+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/session/InMemoryRuntimeSessionRepository.java
@@ -1,14 +1,14 @@
-package com.huawei.ascend.runtime.session.store.memory;
+package com.huawei.ascend.runtime.session;
 
 import com.huawei.ascend.runtime.session.model.Session;
 import com.huawei.ascend.runtime.session.model.SessionKey;
-import com.huawei.ascend.runtime.session.store.SessionStore;
+import com.huawei.ascend.runtime.session.RuntimeSessionRepository;
 
 import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-public final class InMemorySessionStore implements SessionStore {
+public final class InMemoryRuntimeSessionRepository implements RuntimeSessionRepository {
 
     private final ConcurrentHashMap<SessionKey, Session> sessions = new ConcurrentHashMap<>();
