+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/session/RuntimeSessionRepositoryFactory.java
@@ -1,0 +1,5 @@
+package com.huawei.ascend.runtime.session;
+
+public interface RuntimeSessionRepositoryFactory {
+    RuntimeSessionRepository create();
+}
