--- agent-runtime/src/main/java/com/huawei/ascend/runtime/session/config/DefaultSessionStoreFactory.java
@@ -1,21 +1,0 @@
-package com.huawei.ascend.runtime.session.config;
-
-import com.huawei.ascend.runtime.session.store.SessionStore;
-import com.huawei.ascend.runtime.session.store.SessionStoreFactory;
-import com.huawei.ascend.runtime.session.store.memory.InMemorySessionStore;
-
-public final class DefaultSessionStoreFactory implements SessionStoreFactory {
-
-    private final SessionManageProperties properties;
-
-    public DefaultSessionStoreFactory(SessionManageProperties properties) {
-        this.properties = properties;
-    }
-
-    @Override
-    public SessionStore create() {
-        return switch (properties.store().type()) {
-            case MEMORY -> new InMemorySessionStore();
-        };
-    }
-}
