--- agent-runtime/src/main/java/com/huawei/ascend/runtime/session/store/SessionStoreFactory.java
@@ -1,5 +1,0 @@
-package com.huawei.ascend.runtime.session.store;
-
-public interface SessionStoreFactory {
-    SessionStore create();
-}
